requirements/aws_requirements.txt
Django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.6
telethon==1.27.0
scikit-learn==1.2.2
keras==2.12.0
# tensorflow==2.12.0
tensorflow-cpu==2.12.0
cryptography==41.0.3
redis==4.5.5
celery==5.2.7
channels[daphne]==4.0.0
channels-redis==4.0.0
djangorestframework-simplejwt==5.2.2
django-storages==1.13.2
boto3==1.26.147
sentry-sdk==1.25.1
# django-cors-headers==4.0.0
# torch==2.0.0
# torchvision==0.15.0